NEWSPLT-120 Introduced new preference Whitelist.userFeeds to allow UserDefined feeds to be excluded from single-feed usage scenarios
