+++ src/main/java/org/jasig/portlet/newsreader/mvc/portlet/reader/AjaxNewsController.java
@@ -88,7 +88,7 @@ public class AjaxNewsController {
 
         String setName = request.getPreferences().getValue("newsSetName", "default");
         NewsSet set = setCreationService.getNewsSet(setName, request);
-        final List<NewsConfiguration> feeds = AbstractNewsController.filterNonWhitelistedPredefinedConfigurations(request, set.getNewsConfigurations());
+        final List<NewsConfiguration> feeds = AbstractNewsController.filterNonWhitelistedConfigurations(request, set.getNewsConfigurations());
         Collections.sort(feeds);
 
         JSONArray jsonFeeds = new JSONArray();
