+++ src/main/java/org/jasig/portlet/newsreader/service/SharedNewsSetServiceImpl.java
@@ -61,6 +61,7 @@ public class SharedNewsSetServiceImpl implements NewsSetResolvingService {
 		// Persistent set is now loaded but may still need re-initalising since last use.
 		// by adding setId to session, we signal that initialisation has taken place.
         if (session.getAttribute("setId") == null) {
+        	log.debug("re-initalising loaded newsSet "+set.getName());
             @SuppressWarnings("unchecked")
             Set<String> roles = (Set<String>) session.getAttribute("userRoles", PortletSession.PORTLET_SCOPE);
